+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.0.2 - 2019-02-28
+## 1.0.2 - TBC
 ### Changed
 - Readme improvements
+- Improved how custom titles are handled in category elements
+
+### Removed
+- Removed check for tag element
 
 ## 1.0.1 - 2019-02-28
 ### Changed